Update for carousel stories
